CW-3485 Ad Group:: The unallocated budget is shown as 0 if the delivery budget is given less than $1 for Ad group

+++ app/scripts/workflow/controllers/campaign_overview_controller.js
@@ -1112,7 +1112,7 @@ define(['angularAMD', 'common/services/constants_service', 'workflow/services/wo
             $scope.calculateSpendBudget = function (adGroupsData) {
                 var deliveryBudget = $scope.calculateBudget(adGroupsData);
 
-                if (parseInt(deliveryBudget) === 0) {
+                if (parseFloat(deliveryBudget) === 0) {
                     return 0;
                 } else {
                     if (adGroupsData.bookedSpend && adGroupsData.bookedSpend > 0) {
